requirements.txt
ollama==0.4.7
pypdf2==3.0.1
python-docx==1.1.2
streamlit==1.42.0